Move calendar event helpers to events
